quiz-maker/requirements.txt
openai
python-dotenv
mdutils
